# src/routes/index.ts
import { Express, Request, Response } from 'express';
import multer from 'multer';
import { ScreeningOrchestrator } from '../ai/orchestrator';
import { ScreeningRunModel } from '../models/ScreeningRun.model';
import { ScreeningResultModel } from '../models/ScreeningResult.model';
import { AuthController } from '../Controllers/auth.controller';
import { ApplicantsController } from '../Controllers/applicants.controller';
import { ApplicationsController } from '../Controllers/applications.controller';
import { BiasController } from '../Controllers/bias.controller';
import { IngestionController } from '../Controllers/ingestion.controller';
import { JobsController } from '../Controllers/jobs.controller';
import { ResumesController } from '../Controllers/resumes.controller';
import { requireAuth, requireRole } from '../middlewares/auth';

const upload = multer({ storage: multer.memoryStorage(), limits: { fileSize: 10 * 1024 * 1024 } });

export function registerRoutes(app: Express): void {
	// Auth
	app.post('/auth/register', AuthController.register);
	app.post('/auth/login', AuthController.login);
	app.get('/auth/me', requireAuth, (req: Request, res: Response) => {
		res.json({ user: req.user });
	});
	app.post('/auth/logout', (_req: Request, res: Response) => res.status(204).send());
	app.post('/auth/verify', (_req: Request, res: Response) => res.status(501).json({ message: 'Not implemented' }));
	app.post('/auth/resend-code', (_req: Request, res: Response) => res.status(501).json({ message: 'Not implemented' }));

	// Jobs
	app.get('/jobs', JobsController.list);
	app.post('/jobs', requireAuth, requireRole(['recruiter', 'admin']), JobsController.create);
	app.get('/jobs/:jobId', JobsController.getById);
	app.patch('/jobs/:jobId', requireAuth, requireRole(['recruiter', 'admin']), JobsController.update);
	app.post('/jobs/:jobId/activate', requireAuth, requireRole(['recruiter', 'admin']), JobsController.activate);
	app.post('/jobs/:jobId/close', requireAuth, requireRole(['recruiter', 'admin']), JobsController.close);

	// Applicants
	app.get('/applicants', requireAuth, requireRole(['recruiter', 'admin']), ApplicantsController.list);
	app.post('/applicants', requireAuth, ApplicantsController.create);
	app.get('/applicants/:applicantId', requireAuth, ApplicantsController.getById);
	app.patch('/applicants/:applicantId', requireAuth, ApplicantsController.update);

	// Resumes
	app.post('/resumes', requireAuth, upload.single('file'), ResumesController.create);
	app.get('/resumes/:resumeId', requireAuth, ResumesController.getById);
	app.post('/resumes/:resumeId/parse', requireAuth, requireRole(['recruiter', 'admin']), ResumesController.parse);

	// Applications
	app.get('/applications', requireAuth, ApplicationsController.list);
	app.post('/applications', requireAuth, ApplicationsController.create);
	app.patch('/applications/:applicationId', requireAuth, ApplicationsController.update);

	// Screening
	app.post('/screening-runs', async (req: Request, res: Response) => {
		try {
			const { jobId, batchSize, topK, useCache, weightConfig } = req.body || {};
			if (!jobId) {
				res.status(400).json({ message: 'jobId is required' });
				return;
			}
			if (!process.env.GEMINI_API_KEY) {
				res.status(500).json({ message: 'GEMINI_API_KEY not configured' });
				return;
			}
			const orch = new ScreeningOrchestrator(process.env.GEMINI_API_KEY);
			const userId = (req as any).user?.id || 'system';
			const run = await orch.runForJob(jobId, userId, { topK: topK ?? batchSize, weightConfig, useCache });
			res.status(202).json(run);
		} catch (err: any) {
			res.status(500).json({ message: err?.message ?? 'Failed to start screening' });
		}
	});
	app.get('/screening-runs/:screeningRunId', async (req: Request, res: Response) => {
		const run = await ScreeningRunModel.findById(req.params.screeningRunId).lean();
		if (!run) {
			res.status(404).json({ message: 'Not found' });
			return;
		}
		res.json(run);
	});
	app.get('/screening-runs/:screeningRunId/results', async (req: Request, res: Response) => {
		const results = await ScreeningResultModel.find({ screeningRun: req.params.screeningRunId }).sort({ rankPosition: 1 }).lean();
		res.json(results);
	});
	app.get('/screening-results', async (req: Request, res: Response) => {
		const { jobId, top } = req.query as any;
		if (!jobId) {
			res.status(400).json({ message: 'jobId is required' });
			return;
		}
		// find latest completed run for job
		const run = await ScreeningRunModel.findOne({ job: jobId, status: 'completed' }).sort({ completedAt: -1 }).lean();
		if (!run) {
			res.json([]);
			return;
		}
		const limit = Number(top) || 20;
		const results = await ScreeningResultModel.find({ screeningRun: run._id }).sort({ rankPosition: 1 }).limit(limit).lean();
		res.json(results);
	});

	// Bias
	app.get('/bias-audits', requireAuth, requireRole(['recruiter', 'admin']), BiasController.list);
	app.post('/bias-audits/:biasAuditId/dismiss', requireAuth, requireRole(['recruiter', 'admin']), BiasController.dismiss);

	// Ingestion (CSV/XLSX)
	app.post('/ingestion/csv', requireAuth, requireRole(['recruiter', 'admin']), upload.single('file'), IngestionController.ingestCsv);
}